move integration test SocketUtils to kogito-test-utils
