+++ apps-integration-tests/integration-tests-data-index-service/integration-tests-data-index-service-common/src/main/java/org/kie/kogito/index/resources/KogitoServiceRandomPortTestResource.java
@@ -15,8 +15,8 @@
  */
 package org.kie.kogito.index.resources;
 
-import org.kie.kogito.index.utils.SocketUtils;
 import org.kie.kogito.test.resources.TestResource;
+import org.kie.kogito.test.utils.SocketUtils;
 import org.testcontainers.Testcontainers;
 
 public class KogitoServiceRandomPortTestResource implements TestResource {
