+++ apps-integration-tests/integration-tests-jobs-service/integration-tests-jobs-service-common/src/main/java/org/kie/kogito/it/KogitoServiceRandomPortTestResource.java
@@ -15,8 +15,8 @@
  */
 package org.kie.kogito.it;
 
-import org.kie.kogito.it.utils.SocketUtils;
 import org.kie.kogito.test.resources.TestResource;
+import org.kie.kogito.test.utils.SocketUtils;
 import org.testcontainers.Testcontainers;
 
 public class KogitoServiceRandomPortTestResource implements TestResource {
